Improving Smart Wallet landing page -> docs funnel
Took a stab at improving the landing page CTAs for smart wallet by providing a quickstart option via OnchainKit, and docs for more. Structure is more similar to OnchainKit site + docs now, which convert really well.

This is meant to be a jumping-off point rather than a recommendation - going to share this with the team for feedback